refactor(dnd-base): extract spell progression tables to canonical frontend layer
Moves FULL_CASTER_SLOTS, HALF_CASTER_SLOTS, WARLOCK_SLOTS, CANTRIPS_BY_CLASS,
LEVELED_SPELLS_KNOWN_BY_CLASS and PREPARED_SPELL_LEVEL_DIVISOR_BY_CLASS from
creationSpells.ts into entities/dnd-base/spellProgression.ts — the project's
canonical D&D 5e rules layer. Adds cross-reference comment linking the TS file
to its Python counterpart in class_progression.py.
