+++ apps/control-web/src/features/character-sheet/utils/creationSpells.ts
@@ -5,8 +5,13 @@ import {
   getBaseSpellsForClass,
   getSpellAvailabilityClassIds,
   isSameSpellAuthority,
-  resolveSpellSourceClassId,
   resolveSpellByAuthority,
+  resolveSpellSourceClassId,
+  getSlotProgressionForLevel,
+  getMaxUnlockedSpellLevel,
+  getCantripCountForClassLevel,
+  LEVELED_SPELLS_KNOWN_BY_CLASS,
+  PREPARED_SPELL_LEVEL_DIVISOR_BY_CLASS,
 } from "../../../entities/dnd-base";
 import { getModifier } from "./calculations";
 import { getClassCreationConfig } from "../data/classCreation";
@@ -19,152 +24,6 @@ import type {
 type StartingSpellOptionGroup = {
   level: number;
   spells: ReturnType<typeof getBaseSpellsForClass>;
-};
-
-const buildLevelTable = (values: number[]) =>
-  Object.fromEntries(values.map((value, index) => [index + 1, value])) as Record<
-    number,
-    number
-  >;
-
-const buildSlotTable = (rows: Array<Record<number, number>>) =>
-  Object.fromEntries(rows.map((row, index) => [index + 1, row])) as Record<
-    number,
-    Record<number, number>
-  >;
-
-const FULL_CASTER_SLOTS = buildSlotTable([
-  { 1: 2 },
-  { 1: 3 },
-  { 1: 4, 2: 2 },
-  { 1: 4, 2: 3 },
-  { 1: 4, 2: 3, 3: 2 },
-  { 1: 4, 2: 3, 3: 3 },
-  { 1: 4, 2: 3, 3: 3, 4: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 2 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1, 9: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 1, 7: 1, 8: 1, 9: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 2, 7: 1, 8: 1, 9: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 2, 7: 2, 8: 1, 9: 1 },
-]);
-
-const HALF_CASTER_SLOTS = buildSlotTable([
-  {},
-  { 1: 2 },
-  { 1: 3 },
-  { 1: 3 },
-  { 1: 4, 2: 2 },
-  { 1: 4, 2: 2 },
-  { 1: 4, 2: 3 },
-  { 1: 4, 2: 3 },
-  { 1: 4, 2: 3, 3: 2 },
-  { 1: 4, 2: 3, 3: 2 },
-  { 1: 4, 2: 3, 3: 3 },
-  { 1: 4, 2: 3, 3: 3 },
-  { 1: 4, 2: 3, 3: 3, 4: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 2 },
-  { 1: 4, 2: 3, 3: 3, 4: 2 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 },
-  { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 },
-]);
-
-const WARLOCK_SLOTS = buildSlotTable([
-  { 1: 1 },
-  { 1: 2 },
-  { 2: 2 },
-  { 2: 2 },
-  { 3: 2 },
-  { 3: 2 },
-  { 4: 2 },
-  { 4: 2 },
-  { 5: 2 },
-  { 5: 2 },
-  { 5: 3 },
-  { 5: 3 },
-  { 5: 3 },
-  { 5: 3 },
-  { 5: 3 },
-  { 5: 3 },
-  { 5: 4 },
-  { 5: 4 },
-  { 5: 4 },
-  { 5: 4 },
-]);
-
-const SLOT_TABLES_BY_CLASS: Record<string, Record<number, Record<number, number>>> = {
-  bard: FULL_CASTER_SLOTS,
-  cleric: FULL_CASTER_SLOTS,
-  druid: FULL_CASTER_SLOTS,
-  guardian: HALF_CASTER_SLOTS,
-  paladin: HALF_CASTER_SLOTS,
-  ranger: HALF_CASTER_SLOTS,
-  sorcerer: FULL_CASTER_SLOTS,
-  warlock: WARLOCK_SLOTS,
-  wizard: FULL_CASTER_SLOTS,
-};
-
-const CANTRIPS_BY_CLASS = {
-  bard: buildLevelTable([2, 2, 2, 3, 3, 3, 3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4]),
-  cleric: buildLevelTable([3, 3, 3, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5]),
-  druid: buildLevelTable([2, 2, 2, 3, 3, 3, 3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4]),
-  sorcerer: buildLevelTable([4, 4, 4, 5, 5, 5, 5, 5, 5, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6]),
-  warlock: buildLevelTable([2, 2, 2, 3, 3, 3, 3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4]),
-  wizard: buildLevelTable([3, 3, 3, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5]),
-} as const;
-
-const LEVELED_SPELLS_KNOWN_BY_CLASS = {
-  bard: buildLevelTable([4, 5, 6, 7, 8, 9, 10, 11, 12, 14, 15, 15, 16, 18, 19, 19, 20, 22, 22, 22]),
-  ranger: buildLevelTable([0, 2, 3, 3, 4, 4, 5, 5, 6, 6, 7, 7, 8, 8, 9, 9, 10, 10, 11, 11]),
-  sorcerer: buildLevelTable([2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 12, 13, 13, 14, 14, 15, 15, 15, 15]),
-  warlock: buildLevelTable([2, 3, 4, 5, 6, 7, 8, 9, 10, 10, 11, 11, 12, 12, 13, 13, 14, 14, 15, 15]),
-} as const;
-
-const PREPARED_SPELL_LEVEL_DIVISOR_BY_CLASS: Record<string, number> = {
-  cleric: 1,
-  druid: 1,
-  paladin: 2,
-  wizard: 1,
-};
-
-const clampCharacterLevel = (level: number) => Math.max(1, Math.min(20, level));
-
-const getNormalizedSpellProgressionClassId = (className: string) =>
-  resolveSpellSourceClassId(className.trim().toLowerCase());
-
-const getSlotProgressionForLevel = (
-  className: string,
-  level: number
-): Record<number, number> => {
-  const table = SLOT_TABLES_BY_CLASS[getNormalizedSpellProgressionClassId(className)];
-  if (!table) return {};
-  return table[clampCharacterLevel(level)] ?? {};
-};
-
-const getMaxUnlockedSpellLevel = (
-  className: string,
-  level: number,
-  fallbackLevelOneSlots = 0
-) => {
-  const highestSlotLevel = Math.max(
-    0,
-    ...Object.entries(getSlotProgressionForLevel(className, level))
-      .filter(([, count]) => count > 0)
-      .map(([slotLevel]) => Number(slotLevel))
-  );
-
-  if (highestSlotLevel > 0) return highestSlotLevel;
-  return fallbackLevelOneSlots > 0 ? 1 : 0;
 };
 
 const sortSpells = (spells: Spell[]) =>
@@ -356,10 +215,11 @@ export const getFixedStartingSpells = (
         spell !== undefined
     );
 
-const getCantripCountForClassLevel = (className: string, level: number) =>
-  CANTRIPS_BY_CLASS[
-    getNormalizedSpellProgressionClassId(className) as keyof typeof CANTRIPS_BY_CLASS
-  ]?.[clampCharacterLevel(level)];
+// Private helpers used by getLeveledSpellCountForClassLevel and getStartingSpellLimits.
+// The underlying data tables live in entities/dnd-base/spellProgression.ts.
+const clampCharacterLevel = (level: number) => Math.max(1, Math.min(20, level));
+const normalizeSpellProgressionClassId = (className: string) =>
+  resolveSpellSourceClassId(className.trim().toLowerCase());
 
 const getLeveledSpellCountForClassLevel = (
   className: string,
@@ -368,7 +228,7 @@ const getLeveledSpellCountForClassLevel = (
   preparationAbility: keyof CharacterSheet["abilities"] | undefined,
   abilities: CharacterSheet["abilities"]
 ) => {
-  const normalizedClassName = getNormalizedSpellProgressionClassId(className);
+  const normalizedClassName = normalizeSpellProgressionClassId(className);
   const clampedLevel = clampCharacterLevel(level);
 
   if (mode === "spellbook" && normalizedClassName === "wizard") {
